Fix cleanup of temporary response files
We were using the automatic deletion of NamedTemporaryFile()s
to ensure that tarballs created with tegra and module signing
results were getting cleaned up.  This worked OK with Sanic
versions 19.12 and earlier; with later versions, the files
get closed in the sign_handler functions before the
responses are sent, leading to FileNotFound exceptions.

Fix this by disabling auto-deletion of the temp files and
stashing the name in the request ctx field, then deleting
it after the request/response has been completed by installing
a wrapper function for the normal Sanic request handler.

Signed-off-by: Matt Madison <matt@madison.systems>