Refactor handling of workdir, file removals in signing classes
Move construction of a signing object to after the creation
of the temp directory, so we can stash it in the object
and avoid passing it through to the methods.

And since all file removals should be relative to the
working directory, refactor the utils functions to take
that path separately, reducing risk that we'll accidentally
delete files elsewhere due to coding errors.

Signed-off-by: Matt Madison <matt@madison.systems>