rksign: resign without `fit.its`, make uboot.dtb optional
This supports Yocto's kernel-fitimage.bbclass which does the initial
build with UBOOT_MKIMAGE and optionally signs with UBOOT_MKIMAGE_SIGN as
a separate step. In that path, Yocto does not pass the original source
to mkimage and instead uses the '-F' option. If a source is not
provided, mkimage will only re-sign.

Updating the public key in a dtb is also not always necessary, so make
it optional. This could be used to sign additional FIT images containing
scripts or if the public key is populated another way.
kernel-fitimage.bbclass does this conditionally based on whether
UBOOT_DTB_BINARY is defined.

Signed-off-by: Tyler Hall <tyler.hall@lexmark.com>
Signed-off-by: Chad McQuillen <chad.mcquillen@lexmark.com>