tegrasign: check content of scripts before wrapping
The scripts in the 'wrapped' list have historically
required Python 2 be installed as /usr/bin/python, and
need wrapping to invoke the interpreter via 'python2',
since we can't know if 'python' with no version is
available, nor what version it might be.

Now that Python 2 is thoroughly obsolete, and may
not even be available to install, it makes more sense
for users to patch those scripts to be compatible
with Python 3.

To be compatible with both scenarios, check the she-bang
line of each script, and only generate the shell wrapper
for python2 invocation if it ends with 'python' or
'python2'.

Signed-off-by: Matt Madison <matt@madison.systems>